+++ blockstack-sdk/src/main/java/org/blockstack/android/sdk/BlockstackSession.kt
@@ -265,7 +265,6 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
      */
     fun signUserOut() {
         v8userSession.executeVoidFunction("signUserOut", null)
-        v8userSession.release()
         sessionStore.deleteSessionData()
     }
 
@@ -607,13 +606,17 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
                     builder.header(key, headers.getString(key))
                 }
             }
-            blockstackSession.executor.onWorkerThread {
+            blockstackSession.executor.onNetworkThread {
                 val response = httpClient.newCall(builder.build()).execute()
-                blockstackSession.executor.onMainThread {
-                    val r = response.toJSONString()
-                    val v8params = V8Array(v8).push(url).push(r)
-                    v8blockstackAndroid.executeVoidFunction("fetchResolve", v8params)
-                    v8params.release()
+                blockstackSession.executor.onV8Thread {
+                    try {
+                        val r = response.toJSONString()
+                        val v8params = V8Array(v8).push(url).push(r)
+                        v8blockstackAndroid.executeVoidFunction("fetchResolve", v8params)
+                        v8params.release()
+                    } catch (e: Exception) {
+                        Log.d("BlockstackSession", "onfetch", e)
+                    }
                 }
             }
         }
@@ -628,21 +631,30 @@ class BlockstackSession(context: Context? = null, private val config: Blockstack
 
 interface Executor {
     fun onMainThread(function: (Context) -> Unit)
-    fun onWorkerThread(function: suspend () -> Unit)
+    fun onV8Thread(function: () -> Unit)
+    fun onNetworkThread(function: suspend () -> Unit)
 }
 
 class AndroidExecutor(private val ctx: Context) : Executor {
     private val TAG = AndroidExecutor::class.simpleName
     override fun onMainThread(function: (ctx: Context) -> Unit) {
-        launch(UI) { function.invoke(ctx) }
+        launch(UI) {
+            function(ctx)
+        }
     }
 
-    override fun onWorkerThread(function: suspend () -> Unit) {
+    override fun onV8Thread(function: () -> Unit) {
+        launch(UI) {
+            function()
+        }
+    }
+
+    override fun onNetworkThread(function: suspend () -> Unit) {
         async(CommonPool) {
             try {
-                function.invoke()
+                function()
             } catch (e: Exception) {
-                Log.e(TAG, "onWorkerThread", e)
+                Log.e(TAG, "onNetworkThread", e)
             }
 
         }