+++ example-service/src/main/java/org/blockstack/android/Config.kt
@@ -1,9 +1,6 @@
 package org.blockstack.android
 
-val defaultConfig = java.net.URI("https://flamboyant-darwin-d11c17.netlify.com").run {
-    org.blockstack.android.sdk.BlockstackConfig(
-            this,
-            "/redirect",
-            "/manifest.json",
-            kotlin.arrayOf(org.blockstack.android.sdk.Scope.StoreWrite))
-}
+import org.blockstack.android.sdk.toBlockstackConfig
+
+val defaultConfig = "https://flamboyant-darwin-d11c17.netlify.com".toBlockstackConfig(
+        kotlin.arrayOf(org.blockstack.android.sdk.Scope.StoreWrite))